# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sistema de Organización de Eventos — API",
    "version": "1.0.0",
    "description": "API REST del backend del Sistema de Organización de Eventos"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Servidor actual"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    },
    "schemas": {
...
      "Evento": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "nombre": {
            "type": "string"
          },
          "id_tipo_evento": {
            "type": "integer"
          },
          "id_cliente": {
            "type": "integer"
          },
          "fecha_inicio": {
            "type": "string",
            "format": "date"
          },
          "fecha_fin": {
            "type": "string",
            "format": "date"
          },
          "ubicacion": {
            "type": "string"
          },
          "capacidad_estimada": {
            "type": "integer",
            "nullable": true
          },
          "presupuesto_estimado": {
            "type": "string",
            "nullable": true
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "estado": {
            "type": "string",
            "enum": [
              "Planificado",
              "Confirmado",
              "En curso",
              "Finalizado",
              "Cancelado"
            ]
          },
          "dias_anticipacion_recordatorio": {
            "type": "integer"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "EventoProveedor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "id_evento": {
            "type": "integer"
          },
          "id_proveedor": {
            "type": "integer"
          },
          "descripcion_servicio": {
            "type": "string",
            "nullable": true
          },
          "costo_acordado": {
            "type": "string"
          },
          "fecha_servicio": {
            "type": "string",
            "format": "date"
          },
          "hora_inicio": {
            "type": "string",
            "nullable": true
          },
          "hora_fin": {
            "type": "string",
            "nullable": true
          },
          "cantidad": {
            "type": "integer",
            "nullable": true
          },
          "estado_contrato": {
            "type": "string",
            "enum": [
              "Cotizado",
              "Contratado",
              "Pagado",
              "Cancelado"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/eventos/{id}/proveedores": {
      "get": {
        "summary": "Proveedores/servicios contratados para un evento (Cliente solo puede ver el suyo)",
        "tags": [
          "Proveedores"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de contrataciones",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventoProveedor"
                  }
                }
              }
            }
          },
          "403": {
            "description": "No autorizado"
          }
        }
      },
      "post": {
        "summary": "Contrata un proveedor para un evento",
        "tags": [
          "Proveedores"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id_proveedor",
                  "costo_acordado",
                  "fecha_servicio"
                ],
                "properties": {
                  "id_proveedor": {
                    "type": "integer"
                  },
                  "descripcion_servicio": {
                    "type": "string"
                  },
                  "costo_acordado": {
                    "type": "number"
                  },
                  "fecha_servicio": {
                    "type": "string",
                    "format": "date"
                  },
                  "hora_inicio": {
                    "type": "string"
                  },
                  "hora_fin": {
                    "type": "string"
                  },
                  "cantidad": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contratación creada (estado_contrato Cotizado)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventoProveedor"
                }
              }
            }
          },
          "400": {
            "description": "Faltan campos obligatorios"
          },
          "403": {
            "description": "No autorizado (solo Administrador)"
          }
        }
      }
    },
    "/eventos/{id}/proveedores/{contratacionId}": {
      "put": {
        "summary": "Actualiza una contratación (el estado Pagado no se puede setear manualmente)",
        "tags": [
          "Proveedores"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "contratacionId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventoProveedor"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contratación actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventoProveedor"
                }
              }
            }
          },
          "400": {
            "description": "Se intentó setear estado_contrato en Pagado manualmente"
          },
          "403": {
            "description": "No autorizado (solo Administrador)"
          },
          "404": {
            "description": "Contratación no encontrada"
          }
        }
      },
      "delete": {
        "summary": "Cancela una contratación (no se elimina el registro, pasa a estado_contrato Cancelado)",
        "tags": [
          "Proveedores"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "contratacionId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contratación cancelada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventoProveedor"
                }
              }
            }
          },
          "403": {
            "description": "No autorizado (solo Administrador)"
          },
          "404": {
            "description": "Contratación no encontrada"
          }
        }
      }
    },
    "/eventos": {
      "get": {
        "summary": "Lista todos los eventos",
        "tags": [
          "Eventos"
        ],
        "responses": {
          "200": {
            "description": "Lista de eventos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Evento"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "No autorizado (solo Administrador)"
          }
        }
      },
      "post": {
        "summary": "Crea un evento y lo asocia a un Cliente",
        "tags": [
          "Eventos"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nombre",
                  "id_tipo_evento",
                  "id_cliente",
                  "fecha_inicio",
                  "fecha_fin",
                  "ubicacion"
                ],
                "properties": {
                  "nombre": {
                    "type": "string"
                  },
                  "id_tipo_evento": {
                    "type": "integer"
                  },
                  "id_cliente": {
                    "type": "integer"
                  },
                  "fecha_inicio": {
                    "type": "string",
                    "format": "date"
                  },
                  "fecha_fin": {
                    "type": "string",
                    "format": "date"
                  },
                  "ubicacion": {
                    "type": "string"
                  },
                  "capacidad_estimada": {
                    "type": "integer"
                  },
                  "presupuesto_estimado": {
                    "type": "number"
                  },
                  "descripcion": {
                    "type": "string"
                  },
                  "dias_anticipacion_recordatorio": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Evento creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evento"
                }
              }
            }
          },
          "400": {
            "description": "Faltan campos obligatorios"
          },
          "403": {
            "description": "No autorizado (solo Administrador)"
          }
        }
      }
    },
    "/mis-eventos": {
      "get": {
        "summary": "Lista los eventos del Cliente autenticado",
        "tags": [
          "Eventos"
        ],
        "responses": {
          "200": {
            "description": "Lista de eventos del cliente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Evento"
                  }
                }
              }
            }
          },
          "403": {
            "description": "No autorizado (solo Cliente)"
          }
        }
      }
    },
    "/eventos/{id}": {
      "get": {
        "summary": "Detalle de un evento (el Cliente solo puede ver el suyo)",
        "tags": [
          "Eventos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evento"
                }
              }
            }
          },
          "403": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Evento no encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualiza datos del evento",
        "tags": [
          "Eventos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Evento"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Evento actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evento"
                }
              }
            }
          },
          "400": {
            "description": "El evento está cancelado, o se intentó pasar a estado Cancelado por esta vía"
          },
          "403": {
            "description": "No autorizado (solo Administrador)"
          },
          "404": {
            "description": "Evento no encontrado"
          }
        }
      }
    },
    "/eventos/{id}/cancelar": {
      "post": {
        "summary": "Cancela un evento (no se elimina; solo permitido desde Planificado, Confirmado o En curso)",
        "tags": [
          "Eventos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evento cancelado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evento"
                }
              }
            }
          },
          "400": {
            "description": "No se puede cancelar un evento Finalizado o ya Cancelado"
          },
          "403": {
            "description": "No autorizado (solo Administrador)"
          },
          "404": {
            "description": "Evento no encontrado"
          }
        }
      }
    },
    "/eventos/{id}/resumen": {
      "get": {
        "summary": "Resumen agregado del evento (invitados por estado, proveedores contratados, total pagado validado)",
        "tags": [
          "Eventos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resumen del evento"
          },
          "403": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Evento no encontrado"
          }
        }
      }
    },
...
  },
...
}